finished rough version of bandit encounter
finished the structure of the bandit encounter, just need to add proper paragraphs and also potentially fine tuning here and there.

also created a variable for freedDog which is changed to true if the player frees the dog from the bandits.

also moved some less important static code to the bottom of the page to create less clutter